fix(storage): repair state-owner integration hygiene
Tracking: #1710
