+++ crates/astrid-capsule-install/src/checked_tests.rs
@@ -6,7 +6,7 @@ use astrid_core::PrincipalId;
 use astrid_core::dirs::AstridHome;
 use astrid_core::identity::PrincipalUid;
 use astrid_storage::{
-    KvQuotaResolver, PrincipalDirectory, RuntimePrincipalStore, StateOwner,
+    KvQuotaResolver, PrincipalDirectory, RuntimePrincipalStore, StateOwner, StorageError,
     open_runtime_principal_store_with_directory,
 };
 
@@ -40,6 +40,9 @@ fn install_store(home: &AstridHome, principal: &PrincipalId) -> Arc<RuntimePrinc
         Ok(match owner {
             StateOwner::System => None,
             StateOwner::Principal(_) | StateOwner::Fleet(_) => Some(u64::MAX),
+            StateOwner::User(_) => Err(StorageError::Internal(
+                "test quota resolver rejects user StateOwner".to_owned(),
+            ))?,
         })
     });
     let store = Arc::new(
