+++ crates/astrid-capsule/src/context/workspace_branch_tests.rs
@@ -3,7 +3,7 @@ use astrid_core::{
     FirstOwnerClaim, FleetGenesis, FleetIdentity, PrincipalOwnership, UserGenesis, UserIdentity,
 };
 use astrid_crypto::KeyPair;
-use astrid_storage::{KvQuotaResolver, StateOwner};
+use astrid_storage::{KvQuotaResolver, StateOwner, StorageError};
 use chrono::{TimeZone, Utc};
 use uuid::Uuid;
 
@@ -27,6 +27,9 @@ async fn fixture() -> Fixture {
         Ok(match owner {
             StateOwner::System => None,
             StateOwner::Principal(_) | StateOwner::Fleet(_) => Some(u64::MAX),
+            StateOwner::User(_) => Err(StorageError::Internal(
+                "test quota resolver rejects user StateOwner".to_owned(),
+            ))?,
         })
     });
     let store = astrid_storage::open_runtime_principal_store_with_directory(
